Splitted the auth and avatar services
