+++ grpc-server-go/Makefile
@@ -9,4 +9,10 @@ prisma-generate:
 
 gen-prisma: prisma-pull prisma-generate
 
+proto-install:
+	go install google.golang.org/protobuf/cmd/protoc-gen-go@v1.28
+	go install google.golang.org/grpc/cmd/protoc-gen-go-grpc@v1.2
+	export PATH="$PATH:$(go env GOPATH)/bin"
+
+
 gen: gen-proto gen-prisma
